Add variables for a new function along with console.log functions
I have created five new variables for a new function that simulates one round of a game of rock paper scissors. The variable 'playerSelection' takes the users input of 'rock', 'paper', or 'scissors'. The variable 'computerSelection' prints the outcome of the previous fucntion 'getComputerChoice'. The variables 'rock', 'paper', and 'scissors' simply print out their name in a string. The function 'playRound' takes two parameters, 'playerSelection' and 'computerSelection' and plays one round of rock paper scissors by comparing the users input with the computers randomly generated selection.
The console.log functions serve as a method of showing the user what the function is out putting and also showing the users choice, the computers choice, and the overall result of who won.