Add variables for score tracking purposes and console.log messeges
I have added 3 variables 'playerScore', 'computerScore', and 'tieScore'. These variables are used in conjuction with the 'playRound' function in order to keep track of the score during the game.
For every win for the user, 'playerScore' is increased by one point, for every loss for the user 'computerScore' is increased by one point, and for every tie between the user and the computer 'tieScore' is increased by one point.

In the 'game' function, I have added a console.log funtion that displays the final results of the game using the new score tracking variables and along with that I added a messege thanking the user for playing the game and instructing them on how to play the game again.